requirements.txt
beautifulsoup4
pandas
requests
tqdm
google-generativeai
